July26

+++ PopStruc.jl
@@ -1,5 +1,5 @@
 mutable struct Human
-    strains_matrix::Array{Int64,2}
+    strains_matrix::Array{Int8,2}
     Vector_time::Array{Int64,1}
     NumberStrains::Int64
     latenttime::Int64
@@ -28,7 +28,7 @@ function setup_human(h::Array{Human,1})
 end
 
 function create_human()
-    h = Human(zeros(Int64,P.matrix_strain_lines,P.sequence_size),zeros(Int64,P.matrix_strain_lines),0,0,SUSC,UNDEF,0,999,0,0.0,-1,UNDEF,-1,-1,-1,0,-1,0.0)
+    h = Human(zeros(Int8,P.matrix_strain_lines,P.sequence_size),zeros(Int64,P.matrix_strain_lines),0,0,SUSC,UNDEF,0,999,0,0.0,-1,UNDEF,-1,-1,-1,0,-1,0.0)
     #h.strains_matrix = zeros(Int64,P.matrix_strain_lines,P.sequence_size)
     return h
 end
